refactor: divide file about markdown-it plugins
